bitkeeper revision 1.1159.258.100 (4264223aKtYgG3NrSKfbwM7nfkpGcg)
Fix curldebug function in xendomain.c of xfrd. Function takes argument
"size_t buf_num" and uses it as the field width to print, that has to be
an int. While size_t on 32 bit systems is 4bytes (or size of an int) it
is 8bytes on 64bit systems (not size of an int). So this function should
just use "int buf_num". This function is not used in the code anywhere.
This will fix compilation on x86-64.

+++ tools/xfrd/xen_domain.c
@@ -176,7 +176,7 @@ static CURL *curlinit(void){
 
 /** Curl debug function.
  */
-int curldebug(CURL *curl, curl_infotype ty, char *buf, size_t buf_n, void *data){
+int curldebug(CURL *curl, curl_infotype ty, char *buf, int buf_n, void *data){
     // printf("%*s\n", buf_n, buf); /* Does not compile correctly on non 32bit platforms */
     fwrite(data, buf_n, 1, stdout);
     printf("\n");
